[P2P-56] refactor: likeCount, usedCount 타입(Long -> Integer) 변경 및 생성 Dto PhoneNumber 타입 변경

+++ src/main/java/com/prgrms/p2p/domain/place/dto/SummaryPlaceResponse.java
@@ -9,8 +9,8 @@ public class SummaryPlaceResponse {
 
   private Long id;
   private String title;
-  private Long likeCount;
-  private Long usedCount;
+  private Integer likeCount;
+  private Integer usedCount;
   private String category;
   private String thumbnail;
 }
